Picture/Video Views
A View for Displaying Videos has been added. The Video embeds in the
upper left corner of the WebView and leaves quite a bit of whitespace
at the moment. If it’s a YouTube video, the user can click the red Play
icon then the large gray circular Play icon to have the video played
full screen. If it’s a video from Vimeo, the user can click the Play
icon in the lower left corner of the thumbnail and it should play full
screen.

Another View for Displaying multiple Pictures has been added. The user
can use the “Next” and “Previous” Buttons to look through multiple
pictures of each Tour Point.