Rearranged GUI
The GUI was rearranged to resemble the layout of the Android version:

- The building title label was left-aligned above the building picture
in the smTourMapViewController
- A label to display directions to the next tour point was added to the
smTourMapViewController (uses placeholder text)
- A “View Info” button was added to the smTourMapViewController that
segues to the InfoViewController
- An InfoViewController was created with an area to show a picture of
the current building, a description, along with buttons to view more
pictures, view a video, view text, and hear audio. A “Resume Tour”
button was also added.

Note: The Resume Tour, Audio, and Text buttons are not implemented yet.